refactor: restructuring, files and functions relocation
