+++ apps/web/src/components/tasks/CreateTaskForm.tsx
@@ -6,9 +6,9 @@ import { Textarea } from "@repo/ui/components/ui/textarea";
 import { api } from "@repo/backend/convex/_generated/api";
 import { Button } from "@repo/ui/components/ui/button";
 import { zodResolver } from "@hookform/resolvers/zod";
+import { priorities, statuses } from "@/lib/constant";
 import { Input } from "@repo/ui/components/ui/input";
 import { Loader2, Minus, Plus } from "lucide-react";
-import { priorities, statuses } from "@/tasks/data";
 import { createTaskSchema } from "@/lib/validation";
 import { useState, useTransition } from "react";
 import { useMutation } from "convex/react";
