+++ apps/web/src/lib/utils.tsx
@@ -1,61 +1,24 @@
-export const containerVariants = {
-  close: {
-    width: "5rem",
-    transition: {
-      type: "spring",
-      damping: 15,
-      duration: 0.5,
-    },
-  },
-  open: {
-    width: "16rem",
-    transition: {
-      type: "spring",
-      damping: 15,
-      duration: 0.5,
-    },
-  },
-};
+import type { IKanbanColumn, IListColumn } from "@/types/tasks";
 
-export const svgVariants = {
-  close: {
-    rotate: 360,
-  },
-  open: {
-    rotate: 180,
-  },
-};
-
-export const linkVariants = {
-  close: {
-    opacity: 0,
-    transition: {
-      type: "spring",
-      damping: 15,
-      duration: 0.5,
-    },
-  },
-  open: {
-    opacity: 1,
-    transition: {
-      type: "spring",
-      damping: 15,
-      duration: 0.5,
-    },
-  },
-};
-
-export const projectNavVariants = {
-  close: {
-    x: -300,
-    opacity: 0,
-  },
-  open: {
-    x: 0,
-    opacity: 100,
-  },
-}
+import { kanbanColumn, listColumn } from "@/lib/constant";
+import { DEFAULT_CARDS, TableData } from "@/tasks/data";
 
 export function getRandomDate(start: Date, end: Date): string {
   return new Date(start.getTime() + Math.random() * (end.getTime() - start.getTime())).toISOString();
 }
+
+export const getCardByColumn = () => {
+  const cardsByColumn: IKanbanColumn = DEFAULT_CARDS.reduce((acc, card) => {
+    acc[card.column as keyof IKanbanColumn].push(card);
+    return acc;
+  }, kanbanColumn);
+  return cardsByColumn;
+};
+
+export const getTableDataByColumn = () => {
+  const cardsByColumn: IListColumn = TableData.reduce((acc, card) => {
+    acc[card.column as keyof IListColumn].push(card);
+    return acc;
+  }, listColumn);
+  return cardsByColumn;
+};
