+++ apps/web/src/lib/variants.ts
@@ -1,0 +1,57 @@
+export const containerVariants = {
+  close: {
+    width: "5rem",
+    transition: {
+      type: "spring",
+      damping: 15,
+      duration: 0.5,
+    },
+  },
+  open: {
+    width: "16rem",
+    transition: {
+      type: "spring",
+      damping: 15,
+      duration: 0.5,
+    },
+  },
+};
+
+export const svgVariants = {
+  close: {
+    rotate: 360,
+  },
+  open: {
+    rotate: 180,
+  },
+};
+
+export const linkVariants = {
+  close: {
+    opacity: 0,
+    transition: {
+      type: "spring",
+      damping: 15,
+      duration: 0.5,
+    },
+  },
+  open: {
+    opacity: 1,
+    transition: {
+      type: "spring",
+      damping: 15,
+      duration: 0.5,
+    },
+  },
+};
+
+export const projectNavVariants = {
+  close: {
+    x: -300,
+    opacity: 0,
+  },
+  open: {
+    x: 0,
+    opacity: 100,
+  },
+}
